refactor: kafka listener 위치 변경
