+++ src/main/java/KNUCHAT/presentation/MessageHandlingController.java
@@ -1,27 +1,36 @@
 package KNUCHAT.presentation;
 
-import KNUCHAT.domain.ChatMessage;
-import KNUCHAT.application.KafkaService;
+import KNUCHAT.application.ChattingFacade;
+import KNUCHAT.dto.ChatMessage;
+import KNUCHAT.dto.VideoMessage;
+import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @Slf4j
 public class MessageHandlingController {
-    private final KafkaService kafkaService;
-    @MessageMapping("/{id}") // pub
-    @SendTo("/sub/room/{id}") // Sub
-    public ChatMessage testMessage(@DestinationVariable("id") Long id,
-                                   ChatMessage chatMessage){
-        log.info(chatMessage.getMessage());
-        kafkaService.sendMessage(chatMessage);
+    private final ChattingFacade chattingFacade;
 
-        return chatMessage;
+    @MessageMapping("/{id}") // pub
+    public void mapChatMessage(@DestinationVariable Long id,
+                               ChatMessage chatMessage
+    ){
+        log.info("Chat Message : " + chatMessage);
+        chattingFacade.sendMessage(id, chatMessage);
     }
 
+    @KafkaListener(groupId = "video-message-group", topics = "connect-video-call-room", errorHandler = "kafkaListenerErrorHandler")
+    public void consumeVideoMessage(VideoMessage videoMessage){
+
+        log.info("VideoCall Message : " + videoMessage);
+
+        var chatMessage = ChatMessage.from(videoMessage);
+        chattingFacade.sendMessage(chatMessage.roomId(), chatMessage);
+    }
 }
